refactor: settings 관심사 분리
