+++ cypress/e2e/localStorage/localStroage.cy.js
@@ -1,4 +1,4 @@
-import { ERROR_MESSAGE } from "../../../src/settings/settings";
+import { ERROR_MESSAGE } from "../../../src/settings/errorMessages";
 
 describe("로컬 스토리지가 잘 되는지 확인하는 시나리오", () => {
   before(() => {
