+++ src/settings/category.ts
@@ -1,0 +1,16 @@
+export const CATEGORY_ICON = {
+  한식: "./category-korean.png",
+  중식: "./category-chinese.png",
+  일식: "./category-japanese.png",
+  양식: "./category-western.png",
+  아시안: "./category-asian.png",
+  기타: "./category-etc.png",
+} as const;
+export const FOOD_CATEGORY_VALUES = [
+  "한식",
+  "중식",
+  "일식",
+  "아시안",
+  "양식",
+  "기타",
+] as const;
