+++ src/settings/restaurant.ts
@@ -1,19 +1,3 @@
-export const CATEGORY_ICON = {
-  한식: "./category-korean.png",
-  중식: "./category-chinese.png",
-  일식: "./category-japanese.png",
-  양식: "./category-western.png",
-  아시안: "./category-asian.png",
-  기타: "./category-etc.png",
-} as const;
-export const FOOD_CATEGORY_VALUES = [
-  "한식",
-  "중식",
-  "일식",
-  "아시안",
-  "양식",
-  "기타",
-] as const;
 export const RESTAURANT_DISTANCE_VALUES = [
   "5",
   "10",
@@ -26,19 +10,6 @@ export const RESTAURANT_FIELD_LENGTH = {
   name: { min: 1, max: 12 },
   description: { min: 0, max: 300 },
   link: { min: 0, max: 300 },
-} as const;
-
-export const ERROR_MESSAGE = {
-  INVALID_CATEGORY: "존재하지 않는 카테고리 입니다.",
-  INVALID_RESTAURANT_NAME_LENGTH: `음식점 이름은 ${RESTAURANT_FIELD_LENGTH.name.min}글자 이상, ${RESTAURANT_FIELD_LENGTH.name.max}글자 이하만 가능합니다.`,
-  INVALID_RESTAURANT_DISTANCE: "음식점 거리가 유효하지 않습니다.",
-  INVALID_RESTAURANT_DESCRIPTION_LENGTH: `음식점 설명은 ${RESTAURANT_FIELD_LENGTH.description.max}이하만 가능합니다.`,
-  INVALID_RESTAURANT_LINK_LENGTH: `움식점 링크는 ${RESTAURANT_FIELD_LENGTH.link.max}이하만 가능합니다.`,
-  DUPLICATE_RESTAURANT:
-    "이미 동일한 이름의 음식점이 있습니다. 다른 음식점을 입력해주세요.",
-  INVALID_DEFAULT_RESTAURANT:
-    "초기 레스토랑 값이 올바르지 않습니다. 초기 레스토랑 값을 체크해보세요.",
-  NO_RESTAURANT_FOUND: "해당 레스토랑이 없습니다. 이미 지워진것 일수 있어요.",
 } as const;
 
 export const INITIAL_RESTAURANT = [
@@ -133,13 +104,3 @@ export const INITIAL_RESTAURANT = [
     link: "https://bbqchicken.com",
   },
 ] as const;
-export const DICTIONARY = {
-  name: "이름",
-  distance: "거리",
-  all: "모든 음식점",
-  favorite: "자주 가는 음식점",
-  전체: "전체",
-  ...FOOD_CATEGORY_VALUES.reduce((acc, cur) => {
-    return { ...acc, [cur]: cur };
-  }, {} as Record<string, string>),
-} as const;
